tabular doc

+++ fastai/tabular/data.py
@@ -13,7 +13,7 @@ class TabularDataset(DatasetBase):
     def __init__(self, df:DataFrame, dep_var:str, cat_names:OptStrList=None, cont_names:OptStrList=None,
                  stats:OptStats=None, log_output:bool=False):
         if not is_numeric_dtype(df[dep_var]): df[dep_var] = df[dep_var].cat.codes
-        self.y = torch.tensor(df[dep_var].values)
+        self.y = torch.tensor(df[dep_var].values).float()
         if log_output: self.y = torch.log(self.y.float())
         n = len(self.y)
         if cat_names and len(cat_names) >= 1:
